+++ R/sergeant-package.r
@@ -30,7 +30,7 @@
 #' @references \href{https://drill.apache.org/docs/}{Drill documentation}
 #' @docType package
 #' @author Bob Rudis (bob@@rud.is)
-#' @import utils DBI methods bit64 httr jsonlite htmltools
+#' @import utils DBI methods bit64 httr jsonlite htmltools bit64
 #' @importFrom scales comma
 #' @importFrom purrr map map2 map2_df %>%
 #' @importFrom dplyr mutate select left_join bind_cols bind_rows data_frame tbl filter
@@ -38,8 +38,6 @@
 #' @importFrom dplyr db_query_fields src_tbls sql_escape_ident case_when
 #' @importFrom dbplyr build_sql sql_prefix sql_quote src_sql tbl_sql
 #' @importFrom dbplyr win_recycled win_current_group base_win base_agg base_scalar win_over sql
-#' @useDynLib sergeant, .registration = TRUE
-#' @importFrom Rcpp sourceCpp
 NULL
 
 #' sergeant exported operators